chore: split cypress files for less memory (#30354)
(cherry picked from commit 43721f12068c462089c931cc6500610fd2582cab)
