+++ superset-frontend/cypress-base/cypress/e2e/dashboard/nativeFilters.test.ts
@@ -16,15 +16,19 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-import qs from 'querystring';
 import {
   dashboardView,
-  nativeFilters,
-  exploreView,
   dataTestChartName,
+  exploreView,
+  nativeFilters,
 } from 'cypress/support/directories';
-import { SAMPLE_DASHBOARD_1 } from 'cypress/utils/urls';
+import qs from 'querystring';
 
+import {
+  prepareDashboardFilters,
+  SAMPLE_CHART,
+  visitDashboard,
+} from './shared_dashboard_functions';
 import {
   addCountryNameFilter,
   addParentFilterWithValue,
@@ -40,146 +44,15 @@ import {
   fillNativeFilterForm,
   getNativeFilterPlaceholderWithIndex,
   inputNativeFilterDefaultValue,
-  saveNativeFilterSettings,
   nativeFilterTooltips,
+  saveNativeFilterSettings,
+  testItems,
   undoDeleteNativeFilter,
   validateFilterContentOnDashboard,
-  valueNativeFilterOptions,
   validateFilterNameOnDashboard,
-  testItems,
+  valueNativeFilterOptions,
   WORLD_HEALTH_CHARTS,
-  interceptGet,
-  interceptCharts,
-  interceptDatasets,
-  interceptFilterState,
 } from './utils';
-
-const SAMPLE_CHART = { name: 'Most Populated Countries', viz: 'table' };
-
-function visitDashboard(createSample = true) {
-  interceptCharts();
-  interceptGet();
-  interceptDatasets();
-
-  if (createSample) {
-    cy.createSampleDashboards([0]);
-  }
-
-  cy.visit(SAMPLE_DASHBOARD_1);
-  cy.wait('@get');
-  cy.wait('@getCharts');
-  cy.wait('@getDatasets');
-  cy.url().should('contain', 'native_filters_key');
-}
-
-function prepareDashboardFilters(
-  filters: { name: string; column: string; datasetId: number }[],
-) {
-  cy.createSampleDashboards([0]);
-  cy.request({
-    method: 'GET',
-    url: `api/v1/dashboard/1-sample-dashboard`,
-  }).then(res => {
-    const { body } = res;
-    const dashboardId = body.result.id;
-    const allFilters: Record<string, unknown>[] = [];
-    filters.forEach((f, i) => {
-      allFilters.push({
-        id: `NATIVE_FILTER-fLH0pxFQ${i}`,
-        controlValues: {
-          enableEmptyFilter: false,
-          defaultToFirstItem: false,
-          multiSelect: true,
-          searchAllOptions: false,
-          inverseSelection: false,
-        },
-        name: f.name,
-        filterType: 'filter_select',
-        targets: [
-          {
-            datasetId: f.datasetId,
-            column: { name: f.column },
-          },
-        ],
-        defaultDataMask: {
-          extraFormData: {},
-          filterState: {},
-          ownState: {},
-        },
-        cascadeParentIds: [],
-        scope: {
-          rootPath: ['ROOT_ID'],
-          excluded: [],
-        },
-        type: 'NATIVE_FILTER',
-        description: '',
-        chartsInScope: [5],
-        tabsInScope: [],
-      });
-    });
-    if (dashboardId) {
-      const jsonMetadata = {
-        native_filter_configuration: allFilters,
-        timed_refresh_immune_slices: [],
-        expanded_slices: {},
-        refresh_frequency: 0,
-        color_scheme: '',
-        label_colors: {},
-        shared_label_colors: {},
-        color_scheme_domain: [],
-        cross_filters_enabled: false,
-        positions: {
-          DASHBOARD_VERSION_KEY: 'v2',
-          ROOT_ID: { type: 'ROOT', id: 'ROOT_ID', children: ['GRID_ID'] },
-          GRID_ID: {
-            type: 'GRID',
-            id: 'GRID_ID',
-            children: ['ROW-0rHnUz4nMA'],
-            parents: ['ROOT_ID'],
-          },
-          HEADER_ID: {
-            id: 'HEADER_ID',
-            type: 'HEADER',
-            meta: { text: '1 - Sample dashboard' },
-          },
-          'CHART-DF6EfI55F-': {
-            type: 'CHART',
-            id: 'CHART-DF6EfI55F-',
-            children: [],
-            parents: ['ROOT_ID', 'GRID_ID', 'ROW-0rHnUz4nMA'],
-            meta: {
-              width: 4,
-              height: 50,
-              chartId: 5,
-              sliceName: 'Most Populated Countries',
-            },
-          },
-          'ROW-0rHnUz4nMA': {
-            type: 'ROW',
-            id: 'ROW-0rHnUz4nMA',
-            children: ['CHART-DF6EfI55F-'],
-            parents: ['ROOT_ID', 'GRID_ID'],
-            meta: { background: 'BACKGROUND_TRANSPARENT' },
-          },
-        },
-        default_filters: '{}',
-        filter_scopes: {},
-        chart_configuration: {},
-      };
-
-      return cy
-        .request({
-          method: 'PUT',
-          url: `api/v1/dashboard/${dashboardId}`,
-          body: {
-            json_metadata: JSON.stringify(jsonMetadata),
-          },
-        })
-        .then(() => visitDashboard(false));
-    }
-    return cy;
-  });
-}
 
 function selectFilter(index: number) {
   cy.get("[data-test='filter-title-container'] [draggable='true']")
@@ -194,219 +67,6 @@ function closeFilterModal() {
     }
   });
 }
-
-function openVerticalFilterBar() {
-  cy.getBySel('dashboard-filters-panel').should('exist');
-  cy.getBySel('filter-bar__expand-button').click();
-}
-
-function setFilterBarOrientation(orientation: 'vertical' | 'horizontal') {
-  cy.getBySel('filterbar-orientation-icon').click();
-  cy.wait(250);
-  cy.getBySel('dropdown-selectable-icon-submenu')
-    .contains('Orientation of filter bar')
-    .should('exist')
-    .trigger('mouseover');
-
-  if (orientation === 'vertical') {
-    cy.get('.ant-dropdown-menu-item-selected')
-      .contains('Horizontal (Top)')
-      .should('exist');
-    cy.get('.ant-dropdown-menu-item').contains('Vertical (Left)').click();
-    cy.getBySel('dashboard-filters-panel').should('exist');
-  } else {
-    cy.get('.ant-dropdown-menu-item-selected')
-      .contains('Vertical (Left)')
-      .should('exist');
-    cy.get('.ant-dropdown-menu-item').contains('Horizontal (Top)').click();
-    cy.getBySel('loading-indicator').should('exist');
-    cy.getBySel('filter-bar').should('exist');
-    cy.getBySel('dashboard-filters-panel').should('not.exist');
-  }
-}
-
-function openMoreFilters(intercetFilterState = true) {
-  interceptFilterState();
-  cy.getBySel('dropdown-container-btn').click();
-
-  if (intercetFilterState) {
-    cy.wait('@postFilterState');
-  }
-}
-
-describe('Horizontal FilterBar', () => {
-  it('should go from vertical to horizontal and the opposite', () => {
-    visitDashboard();
-    openVerticalFilterBar();
-    setFilterBarOrientation('horizontal');
-    setFilterBarOrientation('vertical');
-  });
-
-  it('should show all default actions in horizontal mode', () => {
-    visitDashboard();
-    openVerticalFilterBar();
-    setFilterBarOrientation('horizontal');
-    cy.getBySel('horizontal-filterbar-empty')
-      .contains('No filters are currently added to this dashboard.')
-      .should('exist');
-    cy.getBySel('filter-bar__create-filter').should('exist');
-    cy.getBySel('filterbar-action-buttons').should('exist');
-  });
-
-  it('should stay in horizontal mode when reloading', () => {
-    visitDashboard();
-    openVerticalFilterBar();
-    setFilterBarOrientation('horizontal');
-    cy.reload();
-    cy.getBySel('dashboard-filters-panel').should('not.exist');
-  });
-
-  it('should show all filters in available space on load', () => {
-    prepareDashboardFilters([
-      { name: 'test_1', column: 'country_name', datasetId: 2 },
-      { name: 'test_2', column: 'country_code', datasetId: 2 },
-      { name: 'test_3', column: 'region', datasetId: 2 },
-    ]);
-    setFilterBarOrientation('horizontal');
-    cy.get('.filter-item-wrapper').should('have.length', 3);
-  });
-
-  it('should show "more filters" on window resizing up and down', () => {
-    prepareDashboardFilters([
-      { name: 'test_1', column: 'country_name', datasetId: 2 },
-      { name: 'test_2', column: 'country_code', datasetId: 2 },
-      { name: 'test_3', column: 'region', datasetId: 2 },
-    ]);
-    setFilterBarOrientation('horizontal');
-
-    cy.getBySel('form-item-value').should('have.length', 3);
-    cy.viewport(768, 1024);
-    cy.getBySel('form-item-value').should('have.length', 0);
-    openMoreFilters(false);
-    cy.getBySel('form-item-value').should('have.length', 3);
-
-    cy.getBySel('filter-bar').click();
-    cy.viewport(1000, 1024);
-    openMoreFilters(false);
-    cy.getBySel('form-item-value').should('have.length', 3);
-
-    cy.getBySel('filter-bar').click();
-    cy.viewport(1300, 1024);
-    cy.getBySel('form-item-value').should('have.length', 3);
-    cy.getBySel('dropdown-container-btn').should('not.exist');
-  });
-
-  it('should show "more filters" and scroll', () => {
-    prepareDashboardFilters([
-      { name: 'test_1', column: 'country_name', datasetId: 2 },
-      { name: 'test_2', column: 'country_code', datasetId: 2 },
-      { name: 'test_3', column: 'region', datasetId: 2 },
-      { name: 'test_4', column: 'year', datasetId: 2 },
-      { name: 'test_5', column: 'country_name', datasetId: 2 },
-      { name: 'test_6', column: 'country_code', datasetId: 2 },
-      { name: 'test_7', column: 'region', datasetId: 2 },
-      { name: 'test_8', column: 'year', datasetId: 2 },
-      { name: 'test_9', column: 'country_name', datasetId: 2 },
-      { name: 'test_10', column: 'country_code', datasetId: 2 },
-      { name: 'test_11', column: 'region', datasetId: 2 },
-      { name: 'test_12', column: 'year', datasetId: 2 },
-    ]);
-    setFilterBarOrientation('horizontal');
-    cy.get('.filter-item-wrapper').should('have.length', 3);
-    openMoreFilters();
-    cy.getBySel('form-item-value').should('have.length', 12);
-    cy.getBySel('filter-control-name').contains('test_10').should('be.visible');
-    cy.getBySel('filter-control-name')
-      .contains('test_12')
-      .should('not.be.visible');
-    cy.get('.ant-popover-inner-content').scrollTo('bottom');
-    cy.getBySel('filter-control-name').contains('test_12').should('be.visible');
-  });
-
-  it('should display newly added filter', () => {
-    visitDashboard();
-    openVerticalFilterBar();
-    setFilterBarOrientation('horizontal');
-
-    enterNativeFilterEditModal(false);
-    addCountryNameFilter();
-    saveNativeFilterSettings([]);
-    validateFilterNameOnDashboard(testItems.topTenChart.filterColumn);
-  });
-
-  it('should spot changes in "more filters" and apply their values', () => {
-    cy.intercept(`/api/v1/chart/data?form_data=**`).as('chart');
-    prepareDashboardFilters([
-      { name: 'test_1', column: 'country_name', datasetId: 2 },
-      { name: 'test_2', column: 'country_code', datasetId: 2 },
-      { name: 'test_3', column: 'region', datasetId: 2 },
-      { name: 'test_4', column: 'year', datasetId: 2 },
-      { name: 'test_5', column: 'country_name', datasetId: 2 },
-      { name: 'test_6', column: 'country_code', datasetId: 2 },
-      { name: 'test_7', column: 'region', datasetId: 2 },
-      { name: 'test_8', column: 'year', datasetId: 2 },
-      { name: 'test_9', column: 'country_name', datasetId: 2 },
-      { name: 'test_10', column: 'country_code', datasetId: 2 },
-      { name: 'test_11', column: 'region', datasetId: 2 },
-      { name: 'test_12', column: 'year', datasetId: 2 },
-    ]);
-    setFilterBarOrientation('horizontal');
-    openMoreFilters();
-    applyNativeFilterValueWithIndex(8, testItems.filterDefaultValue);
-    cy.get(nativeFilters.applyFilter).click({ force: true });
-    cy.wait('@chart');
-    cy.get('.antd5-scroll-number.antd5-badge-count').should(
-      'have.attr',
-      'title',
-      '1',
-    );
-  });
-
-  it('should focus filter and open "more filters" programmatically', () => {
-    prepareDashboardFilters([
-      { name: 'test_1', column: 'country_name', datasetId: 2 },
-      { name: 'test_2', column: 'country_code', datasetId: 2 },
-      { name: 'test_3', column: 'region', datasetId: 2 },
-      { name: 'test_4', column: 'year', datasetId: 2 },
-      { name: 'test_5', column: 'country_name', datasetId: 2 },
-      { name: 'test_6', column: 'country_code', datasetId: 2 },
-      { name: 'test_7', column: 'region', datasetId: 2 },
-      { name: 'test_8', column: 'year', datasetId: 2 },
-      { name: 'test_9', column: 'country_name', datasetId: 2 },
-      { name: 'test_10', column: 'country_code', datasetId: 2 },
-      { name: 'test_11', column: 'region', datasetId: 2 },
-      { name: 'test_12', column: 'year', datasetId: 2 },
-    ]);
-    setFilterBarOrientation('horizontal');
-    openMoreFilters();
-    applyNativeFilterValueWithIndex(8, testItems.filterDefaultValue);
-    cy.get(nativeFilters.applyFilter).click({ force: true });
-    cy.getBySel('slice-header').within(() => {
-      cy.get('.filter-counts').trigger('mouseover');
-    });
-    cy.get('.filterStatusPopover').contains('test_9').click();
-    cy.getBySel('dropdown-content').should('be.visible');
-    cy.get('.ant-select-focused').should('be.visible');
-  });
-
-  it('should show tag count and one plain tag on focus and only count on blur in select ', () => {
-    prepareDashboardFilters([
-      { name: 'test_1', column: 'country_name', datasetId: 2 },
-    ]);
-    setFilterBarOrientation('horizontal');
-    enterNativeFilterEditModal();
-    inputNativeFilterDefaultValue('Albania');
-    cy.get('.ant-select-selection-search-input').clear({ force: true });
-    inputNativeFilterDefaultValue('Algeria', true);
-    saveNativeFilterSettings([SAMPLE_CHART]);
-    cy.getBySel('filter-bar').within(() => {
-      cy.get(nativeFilters.filterItem).contains('Albania').should('be.visible');
-      cy.get(nativeFilters.filterItem).contains('+ 1 ...').should('be.visible');
-      cy.get('.ant-select-selection-search-input').click();
-      cy.get(nativeFilters.filterItem).contains('+ 2 ...').should('be.visible');
-    });
-  });
-});
 
 describe('Native filters', () => {
   describe('Nativefilters tests initial state required', () => {
